Added CSS formatting
To update the navigation, complete the following tasks.

Title: Set the title to "Joey Thinkful." Then, set the title's font size to 52 pixels for mobile and 48 pixels for desktop.
Navigation links: Create four navigation links: Home, About, Portfolio, and Contact. For each, set the font to Montserrat. These links do not need to take the user anywhere.
Background: Set the background color to a dark blue (for example, #0B1C51).
Responsive presentation: Adjust the presentation (spacing of the title and navigation) so that the content looks good in both mobile and desktop views.